requirements.txt
# Core Dependencies (Python 3.13 compatible)
tensorflow>=2.20.0
keras>=3.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Data Science & ML
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.9.0
seaborn>=0.13.0

# Web Framework
Flask>=3.0.0
Werkzeug>=3.0.0

# System Monitoring
psutil>=5.9.0

# Load Testing
locust>=2.15.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# For Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.29.0
notebook>=7.0.0

# Additional ML utilities
scipy>=1.11.0

# Production server (optional)
gunicorn>=21.2.0
